Preprocessor-Premise/requirements.txt
pandas
numpy~=1.21.2
scipy
matplotlib
tables
requests
h2o
logbook
ruamel-yaml
openpyxl
plotly
orca
progressbar
seaborn
tensorflow
now
statsmodels
flask
sklearn